reconfigured barb scale factors

+++ apps/vaporgui/BarbSubtabs.h
@@ -36,8 +36,8 @@ private slots:
     void xDimChanged(int i);
     void yDimChanged(int i);
     void zDimChanged(int i);
-    void lengthChanged(int d);
-    void thicknessChanged(int d);
+    void lengthChanged(double d);
+    void thicknessChanged(double d);
     void recalculateScales();
 
 private:
